+++ routers/voice_to_text.py
@@ -15,7 +15,7 @@ from typing import Annotated
 import re
 import time
 from security.handler import get_current_active_user
-from security.schemas import User
+from security.security_schemas import User
 
 
 router = APIRouter(prefix="/voice_to_text", tags=["Voice to text"])
